+++ app/Widgets/CommunityPosts/CommunityPosts.php
@@ -88,6 +88,7 @@ class CommunityPosts extends Base
         list($origin, $node) = array_values((array)$packet->content);
 
         $view = $this->tpl();
+        $view->assign('contact', \App\Contact::firstOrNew(['id' => $origin]));
 
         $this->rpc(
             'MovimTpl.fill',