+++ app/Widgets/ContactHeader/ContactHeader.php
@@ -2,10 +2,13 @@
 
 namespace App\Widgets\ContactHeader;
 
+use App\Post;
 use App\Widgets\Chats\Chats;
 use App\Widgets\Dialog\Dialog;
 use Movim\Widget\Base;
-
+use Moxl\Xec\Action\Disco\Request;
+use Moxl\Xec\Action\Pubsub\Subscribe;
+use Moxl\Xec\Action\Pubsub\Unsubscribe;
 use Moxl\Xec\Action\Roster\UpdateItem;
 use Moxl\Xec\Payload\Packet;
 
@@ -16,18 +19,48 @@ class ContactHeader extends Base
         $this->registerEvent('roster_additem_handle', 'onUpdate');
         $this->registerEvent('roster_updateitem_handle', 'onUpdate', 'contact');
         $this->registerEvent('roster_removeitem_handle', 'onUpdate', 'contact');
-        $this->registerEvent('vcard_get_handle', 'onVcardReceived', 'contact');
-        $this->registerEvent('vcard4_get_handle', 'onVcardReceived', 'contact');
+        $this->registerEvent('vcard_get_handle', 'onUpdate', 'contact');
+        $this->registerEvent('vcard4_get_handle', 'onUpdate', 'contact');
+        $this->registerEvent('pubsubsubscription_add_handle', 'onSubscription', 'contact');
+        $this->registerEvent('pubsubsubscription_remove_handle', 'onSubscription', 'contact');
+        $this->registerEvent('pubsub_subscribe_errorpresencesubscriptionrequired', 'onSubscriptionPresenceRequired', 'contact');
+        $this->registerEvent('pubsub_subscribe_errorunsupported', 'onSubscriptionUnsupported', 'contact');
+
+        $this->addjs('contactheader.js');
     }
 
     public function onUpdate(Packet $packet)
     {
-        $this->rpc('MovimTpl.fill', '#' . cleanupId($packet->content) . '_contact_header', $this->prepareHeader($packet->content));
+        $this->refreshHeader($packet->content);
     }
 
-    public function onVcardReceived(Packet $packet)
+    public function onSubscription(Packet $packet)
     {
-        $this->rpc('MovimTpl.fill', '#' . cleanupId($packet->content) . '_contact_header', $this->prepareHeader($packet->content));
+        list($jid, $node) = array_values($packet->content);
+
+        if ($node == Post::MICROBLOG_NODE) {
+            $this->refreshHeader($jid);
+        }
+    }
+
+    public function onSubscriptionPresenceRequired(Packet $packet)
+    {
+        list($jid, $node) = array_values($packet->content);
+
+        if ($node == Post::MICROBLOG_NODE) {
+            $this->toast($this->__('communityposts.subscribe_presencerequired'));
+            $this->refreshHeader($jid, disableFollow: true);
+        }
+    }
+
+    public function onSubscriptionUnsupported(Packet $packet)
+    {
+        list($jid, $node) = array_values($packet->content);
+
+        if ($node == Post::MICROBLOG_NODE) {
+            $this->toast($this->__('communityposts.subscribe_unsupported'));
+            $this->refreshHeader($jid, disableFollow: true);
+        }
     }
 
     public function ajaxEditContact($jid)
@@ -38,6 +71,7 @@ class ContactHeader extends Base
 
         $view = $this->tpl();
 
+        $view->assign('jid', $jid);
         $view->assign('contact', $this->me->session->contacts()->where('jid', $jid)->first());
         $view->assign('groups', $this->me->session->contacts()->select('group')->groupBy('group')->pluck('group')->toArray());
 
@@ -53,7 +87,7 @@ class ContactHeader extends Base
             ->request();
     }
 
-    public function ajaxChat($jid)
+    public function ajaxChat(string $jid)
     {
         if (!validateJid($jid)) {
             return;
@@ -65,13 +99,54 @@ class ContactHeader extends Base
         $this->rpc('MovimUtils.redirect', $this->route('chat', $jid));
     }
 
-    public function prepareHeader($jid)
+    public function ajaxSubscribe(string $jid)
     {
-        if (!validateJid($jid)) {
-            return;
-        }
+        if (!validateJid($jid)) return;
+
+        $g = new Subscribe;
+        $g->setTo($jid)
+            ->setNode(Post::MICROBLOG_NODE)
+            ->setFrom($this->me->id)
+            ->request();
+    }
+
+    public function ajaxUnsubscribe(string $jid)
+    {
+        if (!validateJid($jid)) return;
+
+        $g = new Unsubscribe;
+        $g->setTo($jid)
+            ->setNode(Post::MICROBLOG_NODE)
+            ->setFrom($this->me->id)
+            ->request();
+    }
+
+    public function ajaxGetMetadata(string $jid)
+    {
+        $r = new Request;
+        $r->setTo($jid)->setNode(Post::MICROBLOG_NODE)
+            ->request();
+    }
+
+    public function refreshHeader(string $jid, ?bool $disableFollow = false)
+    {
+        $this->rpc(
+            'MovimTpl.fill',
+            '#' . cleanupId($jid) . '_contact_header',
+            $this->prepareHeader($jid, $disableFollow)
+        );
+    }
+
+    public function prepareHeader(string $jid, ?bool $disableFollow = false)
+    {
+        if (!validateJid($jid)) return;
 
         $view = $this->tpl();
+        $view->assign('subscription', $this->me->subscriptions()
+            ->where('server', $jid)
+            ->where('node', Post::MICROBLOG_NODE)
+            ->first());
+        $view->assign('disablefollow', $disableFollow);
         $view->assign('roster', ($this->me->session->contacts()->where('jid', $jid)->first()));
         $view->assign('contact', \App\Contact::firstOrNew(['id' => $jid]));