+++ app/Widgets/ContactHeader/contactheader.js
@@ -1,0 +1,6 @@
+var ContactHeader = {
+    submit: function (button) {
+        button.innerHTML = '<i class="material-symbols spin">progress_activity</i>';
+        button.classList.add('inactive');
+    }
+}